Progress on enemy functionality
Enemy can take damage from projectile
Projectiles now disappear after timer, or on collision with wall/enemy

+++ Assets/Scripts/EnemyScript.cs
@@ -33,7 +33,7 @@ public class EnemyScript : MonoBehaviour
     public void TakeDamage(int damage)
     {
         currentHealth -= damage;
-        healthBar.fillAmount = currentHealth / maxHealth;
+        healthBar.fillAmount = (float) currentHealth / maxHealth;
         if (currentHealth <= 0)
         {
             EnemyDeath();
